Add random movement and relevant component
Added a new system 'random_move' that changes entities' positioning randomly in a grid map. This system writes Position and reads the new component introduced 'MovingRandomly'. This component has been added to the enemy spawner, allowing enemies to move randomly around the map. The random_move system is included in the scheduler to ensure its execution.