Add 'Name' attribute and tooltip feature for characters
This commit adds a 'Name' attribute to the game characters and introduces a tooltip feature. The 'Name' attribute enables each character to have its own name. The tooltip feature provides players with additional information about characters upon hovering. In addition to these, we are handling dead code warning by adding 'allow(dead_code)', so unused functions do not throw any warnings. This can be useful for any future development. Furthermore, there is refactoring done in src/spawner.rs to determine the name, glyph and health of a character ahead of time rather than simultaneously while spawning a monster. The health and name are now included as part of the monster entity. Minor changes were also made in main.rs for the integration of the new features.