Remove collision system, introduce combat system
The collision.rs system was deleted and replaced by the newly created combat.rs system. This change was made to allow characters to attack one another instead of just removing the entity upon collision. Now, entities have a Health component which is decreased when attacked. If the Health reaches 0, the entity is removed. As a result, player_input.rs and random_move.rs were also updated to include the WantsToAttack command when a player or enemy tries to move into a space occupied by an entity. The new WantsToAttack component was added to the components.rs file to support these features.