Add item use functionality and inventory to game
Implemented the ability for the player to use items in the game and view their inventory. A new file for the use_items system was created to handle these actions. Player_input.rs was modified to include "item" destination and "Carried" to check if an item was carried by the player. Also, hotkeys were added to use items directly from the inventory. The components file was expanded with "Carried" and "ActivateItem" components to mark which entity is carrying an item and to activate item use. Changes on hud.rs file allow viewing of carried items on the game interface. Furthermore, changes in the main game loop allow the player to activate the use of items.